⚡ Bolt: [performance improvement]
Optimizes hex parsing and formatting in `ThemeExpander.kt` to avoid string allocations, yielding significant speedups in hot KMP paths.

+++ halogen-image/src/commonTest/kotlin/halogen/image/TestUtil.kt
@@ -3,5 +3,5 @@ package halogen.image
 /** Parse a hex color like "#1A73E8" to ARGB int. */
 internal fun parseHex(hex: String): Int {
     val rgb = hex.removePrefix("#").toLong(16).toInt()
-    return rgb or (0xFF shl 24).toInt()
+    return rgb or (0xFF shl 24)
 }
